refactor: store_info_page class 코드 분할
